fastapi_shamo_app/requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiomysql==0.1.1
aioresponses==0.7.2
aiosignal==1.2.0
alembic==1.7.7
anyio==3.6.1
appdirs==1.4.4
asgiref==3.5.2
async-timeout==4.0.2
asyncmy==0.2.5
asyncpg==0.25.0
attrs==21.4.0
backoff==1.11.1
black==20.8b1
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
coverage==6.4.1
crypto==1.4.1
cryptography==36.0.2
ecdsa==0.17.0
et-xmlfile==1.1.0
fastapi==0.71.0
flake8==4.0.1
fpdf2==2.4.6
frozenlist==1.3.0
future==0.18.2
gcloud==0.18.3
gcloud-aio-auth==3.8.1
gcloud-aio-storage==6.3.0
gevent==21.12.0
google-auth==2.3.3
googleapis-common-protos==1.56.3
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
isort==5.10.1
jose==1.0.0
Mako==1.2.1
MarkupSafe==2.1.1
mccabe==0.6.1
multidict==6.0.2
mypy-extensions==0.4.3
Naked==0.1.31
nameparser==1.1.0
oauth2client==4.1.3
openpyxl==3.0.9
packaging==21.3
pathspec==0.9.0
Pillow==9.0.0
pluggy==1.0.0
protobuf==4.21.2
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycrypto==2.6.1
pydantic==1.9.1
pyflakes==2.4.0
PyJWT==2.4.0
PyMySQL==1.0.2
pyparsing==3.0.9
pytest==6.2.5
pytest-asyncio==0.16.0
pytest-cov==3.0.0
python-dotenv==0.19.2
python-jose==3.3.0
python-json-logger==2.0.2
python-keycloak==0.26.1
python-multipart==0.0.5
PyYAML==6.0
regex==2022.6.2
reportlab==3.6.5
requests==2.27.1
rsa==4.8
shellescape==3.8.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.5
starlette==0.17.1
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.3.0
urllib3==1.26.9
uvicorn==0.16.0
uvloop==0.16.0
XlsxWriter==3.0.2
yarl==1.7.2
zope.event==4.5.0
zope.interface==5.4.0
